Conceal main content in prepartion for backdrop
- A backdrop is the furthest back surface of an app, appearing behind
all other content and components. It's composed of two surfaces: a back
layer (which displays actions and filters) and a front layer (which
displays content). You can use a backdrop to display interactive
information and actions, such as navigation or content filters.

- To avoid showing a division between the top app bar and menu content
that appears on the backdrop, the backdrop has the same color as
the top app bar.

Signed-off-by: Alvin Dizon <alvin.dizon@ghl.com>